+++ engine/losses.py
@@ -8,10 +8,10 @@ def cross_entropy(probs:np.ndarray, target_idx):
     '''
     if probs.ndim == 2:
         row_coords = np.arange(probs.shape[0])
-        p = np.clip(probs[row_coords, target_idx], 1e-12, 1.0)
-        return -np.log(p)
+        p = np.clip(probs[row_coords, target_idx], np.float32(1e-12), np.float32(1.0), dtype=np.float32)
+        return -np.log(p, dtype=np.float32)
     else:
-        return -np.log(probs[target_idx])
+        return -np.log(probs[target_idx], dtype=np.float32)
 
 def cross_entropy_gradient(probs:np.ndarray , target_idx) -> np.ndarray:
     '''
@@ -20,7 +20,7 @@ def cross_entropy_gradient(probs:np.ndarray , target_idx) -> np.ndarray:
     probs_copy = probs.copy()
     if probs.ndim == 2:
         row_coords = np.arange(probs.shape[0])
-        probs_copy[row_coords, target_idx] -= 1
+        probs_copy[row_coords, target_idx] -= np.float32(1.0)
     else:
-        probs_copy[target_idx] = probs_copy[target_idx]  - 1
+        probs_copy[target_idx] = probs_copy[target_idx]  - np.float32(1)
     return probs_copy
